Create committees_and_candidates.json
This file lists the primary committees in our candidates.json file and lists ID's and names the committees have associated themselves with to the FEC. I created it for QA purposes. It's all pretty straight forward, except notice that Michelle Bachmann's House ID is associated with her presidential PAC, and her Presidential ID is associated with her congressional PAC. Also, the primary PAC for Newt Gingrich's 2012 (C00496497) run does not have a correct ID for him. That candidate ID doesn't exist as far as the FEC is concerned.